Aaand... Again

+++ src/ru/fizteh/fivt/students/krivchansky/filemap/WritingUtils.java
@@ -36,7 +36,7 @@ public class WritingUtils {
 
     public void writeOffset(int offset) throws SomethingIsWrongException {
         try {
-            dataFile.writeInt(offset);
+            dataFile.writeByte(offset);
         } catch (IOException e) {
             throw new SomethingIsWrongException("Error acquired while writing to file: " + e.getMessage());
         }
